chore: Update .gitignore, vue.config.js, .eslintrc.js, and package.json
This commit updates the .gitignore file to include the .env file. It also modifies the vue.config.js file to add 'vuex-persist' to the transpileDependencies array. Additionally, it adds a new .eslintrc.js file with the necessary configuration for ESLint. Finally, it updates the package.json file to include the 'axios' dependency and adds the 'babel-eslint' package as a dev dependency.