refactor: Update .gitignore file
This commit updates the .gitignore file to include new directories and files that should be ignored by Git. It removes the .idea and .vscode directories from version control and adds the dist, dist-ssr, coverage, and *.local directories to the ignore list. These changes ensure that unnecessary files and directories are not tracked by Git.